refactor: view-model-controller 폴더 배치
