+++ src/main/kotlin/store/model/entity/OrderItem.kt
@@ -1,3 +1,3 @@
-package store.model
+package store.model.entity
 
 data class OrderItem(val productName: String, var orderQuantity: Int, val price: Int)
