+++ src/main/kotlin/store/model/entity/Promotion.kt
@@ -1,4 +1,4 @@
-package store.model
+package store.model.entity
 
 import java.time.LocalDateTime
